requirements.txt
SQLAlchemy==2.0.29
feedparser==6.0.11
python-dateutil==2.9.0.post0
